requirements.txt
asgiref==3.11.0
crispy-bootstrap5==2025.6
Django==6.0
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-filter==25.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==23.0.0
msgpack==1.1.2
notifications==0.3.2
packaging==25.0
pillow==12.0.0
PyJWT==2.10.1
pytz==2025.2
redis==7.1.0
setuptools==80.9.0
sqlparse==0.5.4
tzdata==2025.3
wheel==0.45.1
whitenoise==6.11.0
dj-database-url==1.0.0
psycopg2==2.9.5

